Move DViews and Logs to their own windows
